requirements.txt
asgiref==3.4.1
Django==3.2.8
django-environ==0.8.1
gunicorn==20.1.0
Pillow==8.4.0
pytz==2021.3
sqlparse==0.4.2
django-recaptcha

